Merge TypeLevel and Erased Flags
Merge TypeLevel and Erased flags - they mean the same thing.

To support this, and also to pass more tests involving typelevel/erased methods,
we need to refactor the way erased decls are handled. We now do the following steps

 - In Typer: no special treatment
 - In PostTyper

    - Replace synthesized implicit arguments to erased parameters
      by default values.
    - Remove inlined code in erased contexts

   The aim of both of these transformations is to reduce code size, while
   not affecting discoverability in IDEs.

 - In CutErasedDecls

    - Make all erased symbols private, using a SymTransformer.
      This is needed so that erased symbols don't poison generation
      of forwarders (in ResolveSuper) or bridges (in Erasure).
    - Replace all code in erased contexts by default values.
      This is needed so that IsInstanceOfChecker works correctly.

 - In Erasure

    - Drop all erased definitions, and check that no references
      to erased symbols remain in the code.

This particular sequence of actions was chosen to address a number of constraints:

 - We can't start to eliminate erased definitions before pattern matching is completed
   as this could hide errors
 - We must eliminate erased members before ResolveSuper
 - It's most convenient to eliminate erased arguments during erasure,
   since that keeps applications and types in sync

Since there is no group of phases between PatternMatcher and ResolveSuper,
we were forced to employ the split approach where we first make the symbol
invisible by making it private and then remove it later, in erasure.